requirements.txt
flask
google-api-python-client==2.36.0
langchain==0.0.169
requests==2.28.2
chromadb==0.3.26
openai==0.27.6
tiktoken==0.4.0
flask-sqlalchemy
pymysql
cryptography
boto3
psycopg2-binary
python-dotenv
